The rest of the highscore methods transferred to the new class <organisation, interface
