+++ highscores
@@ -1,5 +1,5 @@
-1. 1324
-2. 4
+1. 99
+2. 44
 3. 3
 4. 2
 5. 1
